# Conflicts:
#	src/main/java/com/limechain/grandpa/GrandpaService.java
#	src/main/java/com/limechain/grandpa/state/GrandpaSetState.java
#	src/main/java/com/limechain/network/protocol/message/ProtocolMessageBuilder.java
#	src/main/java/com/limechain/rpc/config/CommonConfig.java
#	src/main/java/com/limechain/storage/block/BlockHandler.java
#	src/main/java/com/limechain/sync/warpsync/WarpSyncMachine.java
#	src/main/java/com/limechain/sync/warpsync/WarpSyncState.java
#	src/main/java/com/limechain/sync/warpsync/action/VerifyJustificationAction.java
#	src/test/java/com/limechain/grandpa/GrandpaServiceTest.java

+++ src/main/java/com/limechain/babe/consensus/scale/BabeConsensusMessageReader.java
@@ -25,19 +25,28 @@ public class BabeConsensusMessageReader implements ScaleReader<BabeConsensusMess
         BabeConsensusMessage babeConsensusMessage = new BabeConsensusMessage();
         BabeConsensusMessageFormat format = BabeConsensusMessageFormat.fromFormat(reader.readByte());
         babeConsensusMessage.setFormat(format);
+
         switch (format) {
             case NEXT_EPOCH_DATA -> {
                 List<Authority> authorities = reader.read(new ListReader<>(new AuthorityReader()));
                 byte[] randomness = reader.readUint256();
                 babeConsensusMessage.setNextEpochData(new EpochData(authorities, randomness));
             }
-            case DISABLED_AUTHORITY -> babeConsensusMessage.setDisabledAuthority(reader.readUint32());
+
+            case DISABLED_AUTHORITY ->
+                    babeConsensusMessage.setDisabledAuthority(BigInteger.valueOf(reader.readUint32()));
+
             case NEXT_EPOCH_DESCRIPTOR -> {
-                Pair<BigInteger, BigInteger> constant = new PairReader<>(new UInt64Reader(), new UInt64Reader()).read(reader);
-                BabeEpoch.BabeAllowedSlots secondarySlot = new EnumReader<>(BabeEpoch.BabeAllowedSlots.values()).read(reader);
+                Pair<BigInteger, BigInteger> constant = new PairReader<>(new UInt64Reader(), new UInt64Reader())
+                        .read(reader);
+
+                BabeEpoch.BabeAllowedSlots secondarySlot = new EnumReader<>(BabeEpoch.BabeAllowedSlots.values())
+                        .read(reader);
+
                 babeConsensusMessage.setNextEpochDescriptor(new EpochDescriptor(constant, secondarySlot));
             }
         }
+
         return babeConsensusMessage;
     }
 }
